test: refactor for new structure
