+++ src/Database/Api.hs
@@ -1,4 +1,4 @@
-module Database.Api (editNews, putUser, makeAndFillTables, findUserByLogin, validPassword, validCopyRight, pullAllUsers, findCategoryByLabel, putCategory, editCategory, pullAllCategories, pullImage, putNews, findNewsByTitle, pullAllNews) where
+module Database.Api (editNews, putUser, makeHashPassword, makeAndFillTables, findUserByLogin, validPassword, validCopyRight, pullAllUsers, findCategoryByLabel, putCategory, editCategory, pullAllCategories, pullImage, putNews, findNewsByTitle, pullAllNews) where
 
 import Database.Authorization (validCopyRight, validPassword)
 import Database.Queries.Category (editCategory, findCategoryByLabel, pullAllCategories, putCategory)
@@ -6,3 +6,4 @@ import Database.Queries.Image (pullImage)
 import Database.Queries.News (editNews, findNewsByTitle, pullAllNews, putNews)
 import Database.Queries.User (findUserByLogin, pullAllUsers, putUser)
 import Database.Verb (makeAndFillTables)
+import Database.Crypto (makeHashPassword)
