--- test/Base/CryptoSpec.hs
@@ -1,9 +1,0 @@
-module Base.CryptoSpec (spec) where
-
-import Base.Crypto ()
-import Test.Hspec
-
-spec :: Spec
-spec = do
-  it "nothing to test" $ do
-    (succ 2 :: Int) `shouldBe` (3 :: Int)
