--- test/Base/FillTablesSpec.hs
@@ -1,9 +1,0 @@
-module Base.FillTablesSpec (spec) where
-
-import Base.FillTables ()
-import Test.Hspec
-
-spec :: Spec
-spec = do
-  it "nothing to test" $ do
-    (succ 2 :: Int) `shouldBe` (3 :: Int)
