--- test/Base/TestImageSpec.hs
@@ -1,9 +1,0 @@
-module Base.TestImageSpec (spec) where
-
-import Base.TestImage ()
-import Test.Hspec
-
-spec :: Spec
-spec = do
-  it "nothing to test" $ do
-    (succ 2 :: Int) `shouldBe` (3 :: Int)
