--- test/SchemeSpec.hs
@@ -1,9 +1,0 @@
-module SchemeSpec (spec) where
-
-import Scheme ()
-import Test.Hspec
-
-spec :: Spec
-spec = do
-  it "nothing to test" $ do
-    (succ 2 :: Int) `shouldBe` (3 :: Int)
