--- test/Web/WebTypeSpec.hs
@@ -1,9 +1,0 @@
-module Web.WebTypeSpec (spec) where
-
-import Test.Hspec
-import Web.WebType ()
-
-spec :: Spec
-spec = do
-  it "nothing to test" $ do
-    (succ 2 :: Int) `shouldBe` (3 :: Int)
